% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_bed_stream.R
\name{pull_bed_stream}
\alias{pull_bed_stream}
\title{Pull the stream from the IOBED connection}
\usage{
pull_bed_stream(con, close = TRUE)
}
\arguments{
\item{con}{}
}
\value{

}
\description{
Pull the stream from the IOBED connection
}
