% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_iobed_stream.R
\name{tidy_iobed_stream}
\alias{tidy_iobed_stream}
\title{Tidy version of IOBED stream}
\usage{
tidy_iobed_stream(stream)
}
\arguments{
\item{stream}{(chr) a stream from the IOBED connection}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get a stream pulled by \link{pull_bed_stream} and tidy it in a tibble
}
\note{
If the first row is uncompleted it is ignored and a warning
is signaled (this should not happen!), on the other hand, if the
last row is not completed (almost always!) it is silently ignored.
}
\examples{
sample_stream <- c(
  "2", "7", "8", " ", " ", " ", " ", "2", "2", "5", " ", " ", " ",
  "2", "2", "7", " ", " ", " ", " ", "2", "6", "8", " ", " ", " ",
  " ", "6", "5", "5", "3", "0", " ", " ", " ", " ", " ", "1", "2",
  ":", "4", "4", ":", "1", "4", "/", "2", "0", "0", "\n", "\r", "",
  "2", "7", "8", " ", " ", " ", " ", "2", "2", "5", " ", " ", " ",
  "2", "2", "6", " ", " ", " ", " ", "2", "6", "8", " ", " ", " ",
  " ", "6", "5", "5", "3", "0", " ", " ", " ", " ", " ", "1", "2",
  ":", "4", "4", ":", "1", "4", "/", "2", "0", "1", "\n", "\r", "",
  "2", "7", "8", " ", " ", " ", " ", "2", "2", "5", " ", " ", " "
  )

tidy_iobed_stream(sample_stream)

uscita_stream <- c(
  "2", "7", "9", " ", " ", " ", " ", "2", "2", "5", " ", " ", " ",
  "2", "2", "5", " ", " ", " ", " ", "2", "7", "0", " ", " ", " ",
  " ", "6", "5", "5", "3", "0", " ", " ", " ", " ", " ", "1", "2",
  ":", "5", "0", ":", "2", "t", "i", "p", "o", " ", "u", "s", "c",
  "i", "t", "a", "1", "/", "2", "0", "1", "\n", "\r", "", "2", "7",
  "9", " ", " ", " ", " ", "2", "2", "4", " ", " ", " ", "2", "2",
  "5", " ", " ", " ", " ", "2", "7", "0", " ", " ", " ", " ", "6",
  "5", "5", "3", "0", " ", " ", " ", " ", " ", "1", "2", ":", "5"
)
tidy_iobed_stream(uscita_stream)
}
